+++ Tests/ComposableArchitectureTests/Internal/BaseTCATestCase.swift
@@ -1,4 +1,4 @@
-@_spi(Internals) import ComposableArchitecture
+@_spi(Internals) @_spi(Logging) import ComposableArchitecture
 import XCTest
 
 class BaseTCATestCase: XCTestCase {
@@ -6,5 +6,7 @@ class BaseTCATestCase: XCTestCase {
     super.tearDown()
     XCTAssertEqual(_cancellationCancellables.count, 0, "\(self)")
     _cancellationCancellables.removeAll()
+    Logger.shared.isEnabled = false
+    Logger.shared.clear()
   }
 }
